+++ test/components/test_land_consumption.py
@@ -7,12 +7,12 @@ from approvaltests import verify
 def test_get_basic_table():
     data = pd.DataFrame(
         {
-            'Land Use Object': ['Buildings', 'Agricultural land', 'Natural land', 'Other'],
-            'Land Use Class': ['Commercial', '', '', 'Other land uses'],
-            '% of Consumed Land Area': [100.0, None, None, None],
-            '% of Settled Land Area': [20.0, 10.0, None, 60.0],
-            'Total Land Area [ha]': [20.0, 10.0, 10.0, 60.0],
-            '% of Total Land Area': [20.0, 10.0, 10.0, 60.0],
+            'Land Use Object': ['Buildings', 'Agricultural land', 'Natural land', 'Other', 'Unknown'],
+            'Land Use Class': ['Commercial', '', '', 'Other land uses', 'Unknown land use'],
+            '% of Consumed Land Area': [100.0, None, None, None, None],
+            '% of Settled Land Area': [20.0, 10.0, None, 60.0, None],
+            'Total Land Area [ha]': [20.0, 10.0, 10.0, 10.0, 50.0],
+            '% of Total Land Area': [20.0, 10.0, 10.0, 10.0, 50.0],
         }
     )
 
@@ -24,18 +24,19 @@ def test_get_basic_table():
 def test_get_detailed_table():
     data = pd.DataFrame(
         {
-            'Land Use Object': ['Buildings', 'Buildings', 'Agricultural land', 'Natural land', 'Other'],
+            'Land Use Object': ['Buildings', 'Buildings', 'Agricultural land', 'Natural land', 'Other', 'Unknown'],
             'Land Use Class': [
                 'Commercial',
                 'Institutional',
                 '',
                 '',
                 'Other',
+                'Unknown land use',
             ],
-            '% of Consumed Land Area': [50.0, 50.0, None, None, None],
-            '% of Settled Land Area': [10.0, 10.0, 10.0, None, 60.0],
-            'Total Land Area [ha]': [10.0, 10.0, 10.0, 10.0, 60.0],
-            '% of Total Land Area': [10.0, 10.0, 10.0, 10.0, 60.0],
+            '% of Consumed Land Area': [50.0, 50.0, None, None, None, None],
+            '% of Settled Land Area': [10.0, 10.0, 10.0, None, 60.0, None],
+            'Total Land Area [ha]': [10.0, 10.0, 10.0, 10.0, 10.0, 50.0],
+            '% of Total Land Area': [10.0, 10.0, 10.0, 10.0, 10.0, 50.0],
         }
     )
 
@@ -98,12 +99,12 @@ def test_table_sorted():
 def test_get_treemap():
     data = pd.DataFrame(
         {
-            'Land Use Object': ['Buildings', 'Building', 'Natural land', 'Other'],
-            'Land Use Class': ['Commercial', 'Industrial', '', 'Other'],
-            '% of Consumed Land Area': [66.67, 33.33, None, None],
-            '% of Settled Land Area': [50.0, 25.0, None, 25.0],
-            'Total Land Area [ha]': [2.0, 1.0, 1.0, 1.0],
-            '% of Total Land Area': [40.0, 20.0, 20.0, 20.0],
+            'Land Use Object': ['Buildings', 'Building', 'Natural land', 'Other', 'Unknown'],
+            'Land Use Class': ['Commercial', 'Industrial', '', 'Other', 'Unknown'],
+            '% of Consumed Land Area': [66.67, 33.33, None, None, None],
+            '% of Settled Land Area': [50.0, 25.0, None, 25.0, None],
+            'Total Land Area [ha]': [2.0, 1.0, 1.0, 1.0, 5.0],
+            '% of Total Land Area': [20.0, 10.0, 10.0, 10.0, 50.0],
         }
     )
 
@@ -120,5 +121,7 @@ def test_get_treemap():
         'Land Use Overview/Natural land',
         'Land Use Overview/Other/Other',
         'Land Use Overview/Other',
+        'Land Use Overview/Unknown/Unknown',
+        'Land Use Overview/Unknown',
     ] in fig.data[0].ids
-    assert [5.0, 1.0, 1.0, 2.0, 2.0, 1.0, 1.0, 1.0, 1.0] in fig.data[0].values
+    assert [5.0, 1.0, 1.0, 2.0, 2.0, 1.0, 1.0, 1.0, 1.0, 5.0, 5.0] in fig.data[0].values